rename luxtronic to luxtronik
