--- bundles/org.openhab.binding.luxtronicheatpump/src/main/java/org/openhab/binding/luxtronicheatpump/internal/LuxtronicHeatpumpBindingConstants.java
@@ -1,30 +1,0 @@
-/**
- * Copyright (c) 2010-2020 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.binding.luxtronicheatpump.internal;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.openhab.core.thing.ThingTypeUID;
-
-/**
- * The {@link LuxtronicHeatpumpBindingConstants} class defines common constants, which are
- * used across the whole binding.
- *
- * @author Stefan Giehl - Initial contribution
- */
-@NonNullByDefault
-public class LuxtronicHeatpumpBindingConstants {
-
-    private static final String BINDING_ID = "luxtronicheatpump";
-
-    public static final ThingTypeUID THING_TYPE_HEATPUMP = new ThingTypeUID(BINDING_ID, "heatpump");
-}
