--- bundles/org.openhab.binding.luxtronicheatpump/src/main/java/org/openhab/binding/luxtronicheatpump/internal/LuxtronicHeatpumpHandlerFactory.java
@@ -1,67 +1,0 @@
-/**
- * Copyright (c) 2010-2020 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.binding.luxtronicheatpump.internal;
-
-import static org.openhab.binding.luxtronicheatpump.internal.LuxtronicHeatpumpBindingConstants.*;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.core.thing.Thing;
-import org.openhab.core.thing.ThingTypeUID;
-import org.openhab.core.thing.binding.BaseThingHandlerFactory;
-import org.openhab.core.thing.binding.ThingHandler;
-import org.openhab.core.thing.binding.ThingHandlerFactory;
-import org.osgi.service.component.annotations.Component;
-
-/**
- * The {@link LuxtronicHeatpumpHandlerFactory} is responsible for creating things and thing
- * handlers.
- *
- * @author Stefan Giehl - Initial contribution
- */
-@NonNullByDefault
-@Component(configurationPid = "binding.luxtronicheatpump", service = ThingHandlerFactory.class)
-public class LuxtronicHeatpumpHandlerFactory extends BaseThingHandlerFactory {
-
-    private static final Set<ThingTypeUID> SUPPORTED_THING_TYPES_UIDS = Set.of(THING_TYPE_HEATPUMP);
-    private static final Map<String, LuxtronicHeatpumpHandler> THING_HANDLERS = new HashMap<>();
-
-    @Override
-    public boolean supportsThingType(ThingTypeUID thingTypeUID) {
-        return SUPPORTED_THING_TYPES_UIDS.contains(thingTypeUID);
-    }
-
-    @Override
-    protected @Nullable ThingHandler createHandler(Thing thing) {
-        ThingTypeUID thingTypeUID = thing.getThingTypeUID();
-        LuxtronicHeatpumpHandler thingHandler = null;
-
-        if (THING_TYPE_HEATPUMP.equals(thingTypeUID)) {
-            thingHandler = new LuxtronicHeatpumpHandler(thing);
-        }
-
-        if (thingHandler != null) {
-            THING_HANDLERS.put(thing.getUID().toString(), thingHandler);
-        }
-
-        return thingHandler;
-    }
-
-    public static @Nullable LuxtronicHeatpumpHandler getHandler(String thingUid) {
-        return THING_HANDLERS.get(thingUid);
-    }
-}
