--- bundles/org.openhab.binding.luxtronicheatpump/src/main/resources/OH-INF/binding/binding.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<binding:binding id="luxtronicheatpump" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:binding="https://openhab.org/schemas/binding/v1.0.0"
-	xsi:schemaLocation="https://openhab.org/schemas/binding/v1.0.0 https://openhab.org/schemas/binding-1.0.0.xsd">
-
-	<name>LuxtronicHeatpump Binding</name>
-	<description>This is the binding for a Heatpump with Luxtronic control.</description>
-
-</binding:binding>
