--- bundles/org.openhab.binding.luxtronicheatpump/src/main/resources/OH-INF/i18n/luxtronicheatpump.properties
@@ -1,84 +1,0 @@
-
-channel-type.luxtronicheatpump.SwitchoffX_file_NrX.state.option.1 = Heat pump malfunction
-channel-type.luxtronicheatpump.SwitchoffX_file_NrX.state.option.2 = Facility malfunction
-channel-type.luxtronicheatpump.SwitchoffX_file_NrX.state.option.3 = Second heat generator operating mode
-channel-type.luxtronicheatpump.SwitchoffX_file_NrX.state.option.4 = EVU lock
-channel-type.luxtronicheatpump.SwitchoffX_file_NrX.state.option.5 = Running dew (LW units only)
-channel-type.luxtronicheatpump.SwitchoffX_file_NrX.state.option.6 = Temperature application limit maximum
-channel-type.luxtronicheatpump.SwitchoffX_file_NrX.state.option.7 = Temperature application limit minimum
-channel-type.luxtronicheatpump.SwitchoffX_file_NrX.state.option.8 = Lower operating limit
-channel-type.luxtronicheatpump.SwitchoffX_file_NrX.state.option.9 = No requirement
-
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.701 = Error low pressure - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.702 = low-pressure blocking - reset automatic
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.703 = frost protection - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.704 = Error hot gas - reset in hh:mm
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.705 = motor protection vent. - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.706 = motor protection cp brine - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.707 = coding heat pump - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.708 = Sensor return - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.709 = sensor flow - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.710 = sensor hot gas - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.711 = sensor ambient temp - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.712 = sensor service water - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.713 = sensor heat source in - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.714 = hot gas service water - Reset in hh:mm
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.715 = high pressure off - reset automatic
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.716 = Error high pressure - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.717 = flow sensor heat source - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.718 = ambient maximum - reset automatic
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.719 = ambient minimum - reset automatic
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.720 = min. heat source-temperature - reset automatic in hh:mm
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.721 = Low pressure off - reset automatic
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.722 = difference heating temperature - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.723 = diff. Temp. service water - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.724 = diff. Temp. Defrosting - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.725 = Error service water - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.726 = sensor Mixing circuit 1 - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.727 = brine pressure - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.728 = sensor heat source out - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.729 = Error rotating field - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.730 = screed heating capacity - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.732 = Error cooling - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.733 = Error anode - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.734 = Error anode - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.735 = Sensor external Energysource - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.736 = sensor solar panel - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.737 = sensor solar tank - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.738 = sensor Mixing circuit 2 - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.750 = Sensor return external - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.751 = power supply error
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.752 = power supply/flow error
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.755 = Connection to slave lost - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.756 = Connection to master lost - Please call the installer
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.757 = Error low pressure
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.758 = Error defrosting
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.759 = TDI Message
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.760 = Error defrosting
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.761 = LIN-connection lost
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.762 = intake compressor sensor
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.763 = intake vaporizer sensor
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.764 = sensor compressor-heating
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.765 = Superheat
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.766 = limit of application-CP
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.767 = STB E-Rod
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.770 = Low superheat
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.771 = High superheat
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.776 = limit of application-CP
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.777 = Expansion valve
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.778 = Low pressure sensor
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.779 = High pressure sensor
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.780 = EVI sensor
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.781 = Liquid temp. sensor before EXV
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.782 = Suction gas EVI temp. sensor
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.783 = Communication SEC - Inverter
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.784 = VSS lockdown
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.785 = SEC-Board defective
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.786 = Communication SEC - Inverter
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.787 = VD alert
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.788 = Major VSS fault
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.789 = LIN/Encoding not found
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.790 = Major VSS fault
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.791 = Lost ModBus communication
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.792 = LIN-connection lost
-channel-type.luxtronicheatpump.ERROR_NrX.state.option.793 = Major VSS fault
